net: stmmac: Increase DMA initialization timeout limit
While opening the netdevice, both the PHY and DMA are initialized.

During the initialization procedure of some PHY's, such as the DP83865DVH
in RGMII mode, the receive clock is temporarily turned off. The DMA engine
software reset depends on this clock, and will only complete when it is
there.

I have seen delays of about 2.1-2.2 s. before the DMA engine SW reset
was complete, due to this issue. This patch raises the timeout to 3000 ms.

Signed-off-by: Mischa Jonker <mjonker@synopsys.com>