ARCv2: Signal handler should not assume pt_regs == user_regs_struct
While pt_regs and user_regs_struct used to be binary compatible, this has
changed in the past. Now, with the ARCv2 architectures, this really starts
to cause trouble as ARCv2 uses a completely different pt_regs.

This fixes various issues, such as Ctrl-C causing SIGSEGV, and the ping
command segfaulting.

Signed-off-by: Mischa Jonker <mjonker@synopsys.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>