ARC: [arcfpga] Add DTS for ML509 FPGA Board
This adds a device tree for Xilinx ML-509 board.
This board is considered obsolete, so we do not intend to support it,
however keeping this DTS could come in handy if one would need to do
comparison with SDP.

To use this DTS, configure kernel with fpga_defconfig and edit the path
to  DTS file.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>