ARC: perf: rename ARC_PMU_MAX_HWEVENTS to ARC_PERF_MAX_COUNTERS
We don't really care about amount of countable conditions, we just map
conditions kernel may use in perf subsystem.

As for counters kernel will use as many of them as available in system.
Still it's good to check if its amount matches our expectations.
For now we rely on current maximim value listed in the spec. This is 32
which fits into one machine word ("int").

If number of counters at some point gets higher not only
ARC_PERF_MAX_COUNTERS constant should be bumped but we'll need another
type for usage flags.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>