ARC: [axs101] Fix HDMI frequency divider for v3 motherboard.
On v2 motherboard it was the only clock source for all purposes and for clocking
HDMI phy in particular. It was 25MHz. It lead to a fact that HDMI pixel clock
was 75MHz instead of 74.25MHz required by spec.

Even though on most monitors/TVs we still had normal picture there was a chance
in some corner-cases get problems with image display.

This problem was resolved on v3 motherboard with addition of another clock
source with 27MHz frequency.

Now to support both motherboard at once we check its version and do proper
divider setup.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Signed-off-by: Paul Moers <pmoers@synopsys.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>