xxx: ARC: add compiler barrier to LLSC based cmpxchg (REVISIT)
When auditing cmpxchg call sites, Chuck noted that gcc was optimizing
away some of the desired LDs.

|	do {
|		new = old = *ipi_data_ptr;
|		new |= 1U << msg;
|	} while (cmpxchg(ipi_data_ptr, old, new) != old);

was generating to below

| 8015cef8:	ld         r2,[r4,0]  <-- First LD
| 8015cefc:	bset       r1,r2,r1
|
| 8015cf00:	llock      r3,[r4]  <-- atomic op
| 8015cf04:	brne       r3,r2,8015cf10
| 8015cf08:	scond      r1,[r4]
| 8015cf0c:	bnz        8015cf00
|
| 8015cf10:	brne       r3,r2,8015cf00  <-- Branch doesn't go to orig LD

The solution was to either add volatile to all call sites or simply add
compiler barrier to LLSC based cmpxchg - which although a big sledge
hammer for task - is needed for correctness.

Reported-by: Chuck Jordan <cjordan@synopsys,com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>