ARC: perf: Don't pretend that we can unwind userspace callchain
The current way of userspace unwinding will always fail since we
don't have copy_from_user() to read userspace stack to get the
required regs etc.

So simply record the userspace PC as that is enough to associate a
full kernel call trace with "what was running in userspace at that
time".

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>