ARC: HIGHMEM support
Note that this patch only "enables" support for highmem, there is no
support yet to test it, by parsing a DT mem node etc which tells
where lowmem ends and high mem starts.

Part of problem is that w/o PAE there is no real highmem on ARC:
- 0x8000_0000 to 0xBFFF_FFFF is direct mapped (1G)
- 0xC000_0000 to 0xFFFF_FFFF is peripheral uncached

Only when we get to PAE, does highmem really kick in i.e.
0x1_0000_0000 onwards

However for testing, we cna have multiple mem nodes in DT, where all
despite being in physicla lowmem could be accessed as if in highmem

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>
[vgupta: fixmap code reworked to be simpler/similar to pkmaps]
Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>