ARCv2: Disable SLC if IO coherency unit is absent
Without IO coherency unit for DMA to/from peripherals to work properly
it is required to flush/invalidate SLC in kernel "manually". That will
introduce performance overhead and so SLC existance won't be very
beneficial.

So we simply disable SLC if IOC is missing and then at least we have DMA
exchange without data corruption.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>