axs_nand: get bus width and ECC mode from device tree descrition
As of today we have to deal with both 8- and 16-bit busses on SDP
motherboards v2 and v3 correspondingly. So this option is necessary even
though the plan is to populate this entry in platform init code.

As for ECC mode it's nice to have it for future uses instead of
hard-coded value.

And while at it fix compiler warning when passing parameter improperly
to dev_dbg().

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>