ARC: Anit-optimization continued
With prev commit I intentionally left out inline asm code dealing with
bit fiddling as gcc has no privy into what value gets passed as @nr in
the liline asm.

Then again, the issue is NOT inside the inline asm, but that gcc
optimizers can elide the entire function calling that inline asm based
on large value of @nr - so the prev change is NOT technically complete.

So add clipping of @nr to left over places.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>