ARCv2: spinlock/rwlock: Reset retry delay when starting a new spin-wait cycle
A spin lock could be available momentarily, but the SCOND to actually
acquire it might still fail due to concurrent update from other core(s).
To elide hardware lock, the sequence is retried after a "delay" which is
increased expoenntially to get a nice backoff behaviour.

However, this could cause the delay counter to get to a high value. Thus
when the next spin cycle restarts, reset the counter back to starting
value of 1.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>