[adv7511] Add audio support
This patch adds audio support for the ADV7511 HDMI transmitter
using ALSA SoC.

The code was backported from Analog Devices linux tree from
commit 1770c4a which is available at:
	- https://github.com/analogdevicesinc/linux/commit/1770c4a1e32b158b158fda795bc576cbd1ec1388

The audio can be disabled using menu-config so it is possible
to use only video mode.

Manual CTS mode is not supported so it is used the automatic
CTS mode.

The HDMI mode is automatically started at boot and the audio
(when enabled) registers as a codec into ALSA.

SPDIF DAI format was also added to ASoC as it is required by
adv7511 audio.

Signed-off-by: Jose Abreu <joabreu@synopsys.com>
Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>