PCI: designware: Back-port of minor fixes
dw_pcie_linkretrain() proved to be a not needed step and so the mainline
driver does not have it. dw_pcie_wait_for_link() was created to
centralize
the wait for link routine (part of the pcie-designware core driver).

The link validation logic was moved from our old pcie-snpsdev to the
core driver
in order for other drivers to use it also. For that we changed
dw_pcie_link_up().

Relevant upstream changes limited to DW PCIE driver:
 [1] PCI: designware: Add generic dw_pcie_wait_for_link()
     https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=886bc5ceb5cc3ad4b219502d72b277e3c3255a32

 [2] PCI: designware: Add default link up check if sub-driver doesn't
override
     https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=dac29e6c5460d05774e3e8c4fdf4d6e7bd481fab

[abrodkin: commit message update, link to upstream commits]
Signed-off-by: Joao Pinto <jpinto@synopsys.com>
Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>