mmc: dw_mmc: Take care about card_number field in CMD register
CMD[20:16] or "card_number" bits specify сard number in use, it represents
physical slot number of card being accessed.

It was a pure luck that default value thee is 0 and so for slot #0
everything used to work properly. But to use other slots we do need
to set correct value here.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>