ARCv3: link above 4GB: handle 64-bit literals in final link
[Fix link errors with -mcmodel=large and LINK_BASE=0xFFFF_0000_0000 #1]

In ARC64, instructions such as
   LDL d, [@symbol] or J <@symbol>

won't work if @symbol is an absolute address 64-bits wide since the max
size of instruction with a literal, can't be more than 8 bytes.

When kernel is linked at high address, say 0xFF_0000_0000, the absolute
addresses end up as 64-bit literals which can't be encoded in instructions
such as  LDL d, [@symbol] or  J <@symbol> since maz size of instruction
in ARC64 with opcodes + literals is still 64-bits.

|  LD      .tmp_vmlinux1
| arch/arc/kernel/entry-arcv3.o: in function `EV_TLBProtV':
| (.text+0x474): relocation truncated to fit: R_ARC_32_ME against symbol `ret_from_exception' defined in .text section in arch/arc/kernel/entry-arcv3.o
| make[1]: *** [/home/vineetg/arc/v2-kernel/Makefile:1078: vmlinux] Error 1

|  LD      .tmp_vmlinux1
| arch/arc/kernel/entry-arcv3.o: in function `ret_from_fork':
| (.text+0x51c): relocation truncated to fit: R_ARC_32_ME against symbol `schedule_tail' defined in .text section in kernel/sched/core.o

So the solution is to
 - split up the 64-bit limms to low/high and stitch them in code
   (this can be expensive)
 - Jumps need to be converted into PC relative branches

Note that with-mcmodel=large but canonical LINK_BASE=0x8000_0000 these
errors don't happen since all the final addresses resolve to 32-bits
only. The issue only haooens when addresses are wider than 32-bits such
as when linking the kernel binary above 0xFFFF_FFFF

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>