ARC: Add option to control SCM
Shared Cache Memory (SCM) is L2 cache used in ARCv3.

Sometimes we don't want to have it enabled (to test performance
change), so need to have an option to control it.

Please note that this option works only for ARCv3 SCM, because in
ARCv2 L2 cache is enabled by default and cannot be disabled in such
easy way.

Signed-off-by: Vladimir Isaev <isaev@synopsys.com>