ARC: mm: Move kernel to 0xffff_0000_0000_0000
In order to provide more memory to userspace we need to
move kernel to some high address. Our MMU architecture
allows us to configure two root tables, one for 0x0 and one
for 0xffff_0000_0000_0000 (for MMUv48, 16/16 case). It is natural
to use first one for userspace and second one for kernel space, so
we will have kernel mapped in MMU without remapping it for
each user process.

After this change link addr becomes != to phy so we need to enable
MMU as early as possible, to have correct addresses for global variables
and jumps.

This patch fixes some phy != link problems, adds early MMU configuration
and moves kernel to high address.

Signed-off-by: Vladimir Isaev <isaev@synopsys.com>