ARC: mm: Add loader build target
For some tools/simulators/debuggers we have to provide ELF file
with physical entry point and load address. It was not a problem
because for 1:1 we can use vmlinux.

vmlinux contains virtual addresses so it cannot be used when virt != phy.
For such cases loader target can be used. loader is actually simple ELF file
with load and entry point addresses converted to phy space and vmlinux.bin
inside.

Signed-off-by: Vladimir Isaev <isaev@synopsys.com>