ARCv3: mm: Use some 64-bit ops on pgd/pud/pmd/pte entries
This is preparing for MMU48 which has 64-bit pgd/pud/pmd/pte entries.

As of today, although the entries are 64-bit, only lower 32-bits are
valid (since evetythign is still mapped under 4 GB). So the 32-bit
ALU ops in TLB Miss asm code are sufficient.

MMU48 has PGD_SHIFT 39, so we need 64-bit ALU operations, at least the
vaddr shifting to find indexes into various tables. This is needed even
when values are 32-bits, just as a stepping stone.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>