ARC: Early map 1Gb for uboot args
Before this patch we mapped only PAGE_OFFSET -> _end at very early stage.
It is enough for kernel to work, but it is possible that we need to access
memory outside of kernel to read U-boot arguments for example.

Since we have no information about how much memory do we have at the early
boot, we have to map some reasonable amount of memory for U-boot args or some
other stuff.

Signed-off-by: Vladimir Isaev <isaev@synopsys.com>