ARC: Add page attributes to the memory attribute register
MMU memory attribute register has to be the same for all cores.

Core 0 tunes this register in 2 places. Second time it adds
MEMATTR_IDX_UNCACHED and MEMATTR_VOLATILE properties. But others
cores(!=0) in the SMP configuration setuped only MEMATTR_NORMAL to this
register. This fix lets to have the same value in this register for all
cores.