ARCv3: HS5x: set explisit rule to use generic atomic64_* funcs
The configuration with enabled ARC_HAS_LL64 and ARC_HAS_LLSC
options cause build error. Add explicit rule to use generic
atomic64_* funcs for HS5x.

It is a temporary solution, when support to llockd/scondd
inctructions will be added to the toolchain it will be
possible to add optimized implementation.

Signed-off-by: Pavel Kozlov <pavel.kozlov@synopsys.com>