ARCv3: add new version of 32-bit arch_atomic_* with atld.<op> instrs
Add new implementation of 32-bit atomic funcs with atld.<op>
instructions that can be used if ARC_HAS_ATLD enabled. Also add
possibility to use ARC_HAS_ATLD with or without ARC_HAS_LLSC. It will
allow to use ATLD variant in case when LLSC is not desirable
(because of llock/scond livelock issue) or combine and use ATLD for
atomic functions and LLSC for cmpxchg.

Signed-off-by: Pavel Kozlov <pavel.kozlov@synopsys.com>