ARCv3: HS5x: fix compiler option for the ARC_HAS_LL64 config variable
Current scheme with -mcpu=hs5x and -mno-ll64 if CONFIG_ARC_HAS_LL64 not
set is incorrect. We expect that option -mcpu=hs5x tells the compiler
to use duals loads/stores by default, but actually it is not true.
By default -mcpu=hs5x doesn't enable use of 64-bit dual loads/stores and
to enable them we have to add -mll64 option. It is reverse logic to one
used for ARCv2 (hs38).
So, add extra conditions to implement new scheme with -mcpu=hs5x and
-mll64 (for ll64) for ARCv3 32-bit architecture configs.

Signed-off-by: Pavel Kozlov <pavel.kozlov@synopsys.com>